requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
openai==1.3.0 